Rename sandbox util file to apps
